anti-phishing-ext-main/requirements.txt
absl-py==2.1.0
asttokens==2.4.1
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cssselect==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
dnspython==2.6.1
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
fonttools==4.51.0
gast==0.5.4
google==3.0.0
google-pasta==0.2.0
googlesearch-python==1.2.4
grpcio==1.62.2
h5py==3.11.0
idna==3.7
ipykernel==6.29.4
ipython==8.23.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
libclang==18.1.1
lxml==5.2.2
lxml_html_clean==0.1.1
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
nest-asyncio==1.6.0
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==6.0.0
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.0.0
readability==0.3.1
readability-lxml==0.8.1
requests==2.32.3
requests-file==2.1.0
rich==13.7.1
scikit-learn==1.3.2
scipy==1.14.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
termcolor==2.4.0
threadpoolctl==3.5.0
tld==0.13
tldextract==5.1.2
tornado==6.4
traitlets==5.14.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
wcwidth==0.2.13
Werkzeug==3.0.3
whois==1.20240129.2
wrapt==1.16.0
